chore: Split server utils into individual files
